# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayPass API
  description: |
    PayPass is a comprehensive payment platform providing secure, scalable payment processing solutions.
    
    ## Features
    - Payment Processing
    - Wallet Management  
    - User Authentication
    - Enterprise Billing
    - Fraud Detection
    - Compliance Management
    
    ## Authentication
    PayPass API uses API key authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```
    
    ## Rate Limiting
    API requests are rate limited based on your subscription plan:
    - Basic: 100 requests/hour
    - Premium: 1000 requests/hour  
    - Enterprise: 10000 requests/hour
    
    ## Webhooks
    PayPass can send webhook notifications for various events. Configure webhooks in your dashboard.
    
  version: 2.1.0
  contact:
    name: PayPass API Support
    email: api-support@paypass.com
    url: https://docs.paypass.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://paypass.com/terms
servers:
  - url: https://api.paypass.com/v1
    description: Production server
  - url: https://api-staging.paypass.com/v1
    description: Staging server
  - url: https://api-dev.paypass.com/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Create a user
      description: Register a new user account
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    get:
      tags:
        - Users
      summary: List users
      description: Retrieve a list of users
      operationId: listUsers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Enterprise Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - first_name
        - last_name
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        first_name:
          type: string
          maxLength: 50
          description: First name
        last_name:
          type: string
          maxLength: 50
          description: Last name
        phone:
          type: string
          description: Phone number
        country:
          type: string
          description: Country code
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: Email address
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        phone:
          type: string
          description: Phone number
        country:
          type: string
          description: Country
        status:
          type: string
          enum: [active, inactive, suspended]
          description: User status
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    # Wallet Schemas
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token
        token_type:
          type: string
          example: Bearer
          description: Token type
        expires_in:
          type: integer
          description: Token expiry in seconds
        user:
          $ref: '#/components/schemas/User'

    # Enterprise Schemas
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        total:
          type: integer
          description: Total items
        has_more:
          type: boolean
          description: More items available
...